week3-docker-sqlalchemy/day5/ecommerce_pipeline/requirements.txt
SQLAlchemy>=1.4
psycopg2-binary>=2.9
python-dotenv>=0.20
pandas==2.3.3